backend/brain-engine/requirements.txt
redis
numpy
pandas
scikit-learn
joblib
fastapi
uvicorn[standard]
python-dotenv